examples/e2e_app/requirements.txt
azure-functions
azure-functions-db-python[all]
